backend_clean/requirements.txt
# ===========================================
# NEONA CHAT DEMO - Python Dependencies
# ===========================================
# Install these dependencies with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0  # MongoDB driver

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.0

# Azure Services
azure-cognitiveservices-speech==1.34.0
openai==1.3.8

# Audio Processing
pydub==0.25.1
wave

# Date/Time
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Validation
pydantic==2.5.0

# CORS
fastapi-cors

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1